refactor: ProductManager에서 Product 리스트 아이템 로직 ProductItem으로 분리
